Rewrite test_communicator_geo
